Fix response handler to allow string returns
Use case: we use the `pushResponseHandler` method to define
a custom response handler that does some string replacements
on the final response. This means we return a string instead
of a response.
Since the value returned by the function is casted to a string
anyways, it should be able to return a string directly without
any issues.